code analysis fixing

+++ BingoUI/Source/CustomWidgets/CreditosLabel.cpp
@@ -14,6 +14,6 @@ void CreditosLabel::Update()
     player_credits << "$ " << PlayerRef.creditsLeft();
     std::string creditsString = player_credits.str();
     SDL_Rect size = Graphics::MeasureText(creditsString);
-    SetPosition({static_cast<int>(1120 - size.w / 2), 20});
+    SetPosition({1120 - size.w / 2, 20});
     SetText(creditsString);
 }
